:bug: Fix undefined references in GCC
GCC needs a call to throw to force it to link in the exception handling
code. For code size, this a great, but the linker then complains that
the -Wl,wrap= commands are undefined. Those undefined symbols result in
additional functions like _exit not being defined causing a few
linker error messages.

- :arrow_up: 0.0.2