:zap: Optimize (#26)
* :zap: Add the folowing optimizations (1)

- :recycle: Simplify `restore_cpu_core` assembly a bit
- :zap: Unwind becomes a jump table

* :zap: Jump table working. May have bugs

* :zap: [[gnu::always_inline]] Many functions

- :zap: Carve out a branch for decoding uleb128 call-site information,
  the most common and encouraged encoding.
- :bug: Add `lr` and `pc` to clobber list to ensure that the compiler
  does not attempt to use the `lr` register to hold the pointer to the
  virtual_cpu structure
- :zap: Move index entry definitions to the header

* :bug: Get multi levels demo to work

* :zap: Many fixes and improvements

* :construction: Turn on destructors (currently broken)

* About this code

I was trying to make a version of the `enter_function()` function that
has less branches. Trading space for speed.

* :zap: Add enter_function2()

`enter_function2()` will be a replacement for `enter_function()`. This
function correctly decodes the call sites using their encoding values
and decodes the action table and type table correctly as well. Although
the type table parsing is still assumed to be pcrel. This should be
fixed at some point.

* :poop: unwind_frame2

The idea of unwind_frame2 was to evaluate the instructions in place vs
putting them into order in the instruction_t. The problem is word
boundaries which requires logic to determine if we are at the end of the
instructions.

* :fire: Remove some uneeded unwind code

* :zap: Cache & Use the personality pointer

* :wastebasket: Remove old unneeded code

* :zap: Remove usage of personality in create_instructions_from_entry

* :memo: Add todo w/ github issues everywhere in the code

- Remove a few `#if 0` code
- Slightly fix up issues with CI