:sparkles: Support polymorphic catch
Exception algorithm will now check if a parent type exists within the
thrown object that can be caught by a catch block. There is a limit of
up to 12 relationships that can exist for a thrown object. Any greater
than that and the exception mechanism will call `std::terminate()`.
A relationship is a connection from parent to child object. With virtual
inheritance, an additional relationship is made for each virtual parent
inheritance.

- :bug: Fix `read_sleb128()`, incorrectly sign extends values.
- :bug: Fix `get_action_decoder()` next action record offset.

Resolves #6