:bug: (patch) Wrap _Unwind_Resume (#78)
When using LTO, the compiler may emit a `_Unwind_Resume` call rather
than a `__cxa_end_cleanup` at the end of a landing pad. Calling GCC's
unwinder while the estell unwinder is running will be UB. This wraps
that API to call our `__cxa_end_cleanup`. The pointer to the
current exception object passed to _Unwind_Resume, is ignored.