:zap: (minor) Add nearpoint search support (#92)
Nearpoint search reduces the time required to perform an exception entry
lookup by grouping transparent functions by their unwind information and
sorting all groups of functions by size from largest to smallest. This
ordering allows for a lookup table to be built that can provide
information of where to find an exception entry. The goal for valid
nearpoint tables is to ensure that for all inputs, every output is
within 8 entries of the desired entry.

The current format of the nearpoint table and descriptors is
experimental and thus may change between new updates.